fix types for VS2005
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/trunk@1063 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ common/arc.c
@@ -228,7 +228,7 @@ const OEMCHAR	*p;
 	if (p == NULL) {
 		return(plainfile_regist(file_open(fname)));
 	}
-	len = p - fname;
+	len = (UINT)(p - fname);
 	if (len >= NELEMENTS(path)) {
 		return(NULL);
 	}
@@ -263,7 +263,7 @@ const OEMCHAR	*p;
 	if (p == NULL) {
 		return(file_attr(fname));
 	}
-	len = p - fname;
+	len = (UINT)(p - fname);
 	if (len >= NELEMENTS(path)) {
 		return(-1);
 	}
